make colours global
